Adding additional process for Playexisting and change of distance
In playexisting function part, at the begining and ending of writting, a rising process and a falling process was added. This is to prevent the robot  from creating additional strokes during  rail movement.  The rail moving distance of  sanlianxie function change because we remeasure the distance and get a more accurate value.