Adding function Play_Text_updown
Play_Text_updown is the upgraded version of last Play_Text. Previous function have a problems in moving down at the begining of writing. Sometimes this could lead to a delayed movement and results in a wrong output image. In this version, down movement was replaced by a direct instruction on motor (ID:2) and this seems to solve this problem. Play_Text in this version was the one without adding moving down at the begining and moving out at the end. some PlayText in sanlianxie was replaced by PlayText_updown.